+++ src/components/carousel/index.d.ts
@@ -33,7 +33,7 @@ export function scrollPrev(element: HTMLElement | Element): void
 export function scrollNext(element: HTMLElement | Element): void
 export function scrollTo(element: HTMLElement | Element, selected?: number): void
 export function getItemCount(element: HTMLElement | Element, scrollWidth?: number, mathFloor?: boolean): number
-export function observeCarousel(element: HTMLElement | Element, options?: ObserveCarouselOptions): void
+export function observeCarousel(element: HTMLElement | Element, options?: ObserveCarouselOptions): IntersectionObserver
 export function scrollCarousel(element: HTMLElement | Element, options?: ScrollCarouselOptions): void
 export function paginationCarousel(element: HTMLElement | Element, options?: PaginationCarouselOptions): void
 export function autoplayCarousel(element: HTMLElement | Element, options?: AutoplayCarouselOptions): void