+++ src/components/details/index.d.ts
@@ -1,9 +1,5 @@
-export interface DefaultOptions {
+export interface ToggleDetailsOptions {
     selector?: string
-    summarySelector?: string
 }
 
-export const defaultOptions: DefaultOptions
-export function showDetails(selector: HTMLInputElement | HTMLElement, options?: DefaultOptions): Promise<void>
-export function closeDetails(selector: HTMLInputElement | HTMLElement, options?: DefaultOptions): Promise<void>
-export function toggleDetails(selector: HTMLInputElement | HTMLElement, options?: DefaultOptions): Promise<void>
+export function toggleDetails(element: HTMLInputElement, options?: ToggleDetailsOptions): void