+++ src/components/form/index.d.ts
@@ -1,29 +1,11 @@
+import { validateField, type ValidateFieldOptions } from "../field";
+
 export interface ValidateFormOptions {
-    validateSelectors?: string
+    validateSelector?: string
     validateOptions?: ValidateFieldOptions
     validateField?: typeof validateField
     scrollOptions?: ScrollIntoViewOptions
     submitterLoadingAttribute?: string
 }
 
-export interface ValidateFieldOptions {
-    validate?: boolean
-    validationMessage?: string
-    selector?: string
-    ignoreMatch?: RegExp
-    validitySelector?: string
-    infoParentSelector?: string
-    infoSelector?: string
-    infoContent?: string
-    endParentSelector?: string
-    endSelector?: string
-    endContent?: string
-    validAttribute?: string
-    validIcon?: string | null
-    invalidAttribute?: string
-    invalidIcon?: string
-    activeAttribute?: string
-}
-
 export function validateForm(event: Event | SubmitEvent, options?: ValidateFormOptions): void
-export function validateField(element: HTMLElement | SubmitEvent, options?: ValidateFieldOptions): void