+++ src/supports.js
@@ -1,6 +1,16 @@
 /**
  * @type {boolean}
  */
+export const supportsAnchor = CSS.supports('anchor-name', '--')
+
+/**
+ * @type {boolean}
+ */
+export const supportsAnchoredContainer = CSS.supports('container-type: anchored')
+
+/**
+ * @type {boolean}
+ */
 export const supportsTimelineTrigger = CSS.supports('timeline-trigger-name: --v')
 
 /**
@@ -17,3 +27,8 @@ export const supportsScrollInitialTarget = CSS.supports('scroll-initial-target',
  * @type {boolean}
  */
 export const supportsAnimationTimeline = CSS.supports('animation-timeline: scroll()')
+
+/**
+ * @type {boolean}
+ */
+export const supportsScrollSnapEvents = ('onscrollsnapchanging' in window)